Release 1.0.0 with all backwards-compatibility dropped
My reason to touch this is the awkward, outdated version number.

I believe the .php entry point is not needed any more. The extension.json
already contains the same "MessagesDirs" setting. I had a look at the
mentioned mergeMessageFileList.php and believe it does not need the
global $wgMessagesDirs variable any more. But I'm not 100% sure about
this.

Depends-On: Iebbc3849b638c16daa6fe8debb5fa9d81a811c1a
Change-Id: I6b407bf9cc73d9d2ac3db04f23b9759a7510536c