Switch (back) to the MediaWiki CodeSniffer version 16
This component barely contains any PHP code anyway. None of the reasons
why the Wikibase CodeSniffer standard was created is relevant here.
(Except one, now visible as an <exclude …> in the .phpcs.xml file.)
It makes maintenance a little easier if codebases that easily can do
so use the MediaWiki standard directly.

Bug: T188086
Change-Id: I1d0efd9247fe76bc347039075e92ec267c6fa4dd