Remove @embed for badge images
This stylesheet is loaded on all page views from the BeforePageDisplay
hook. That means all possible badge icons and variations were blockingly
downloaded before any part of the skin layout or article text can be
visually rendered – of any page, including those without any badges.

The space where the icons render is reserved, which means loading
these on-demand from CSS normally would not causes any reflow.

The dimensions of the icons are fairly small and typically outside
the viewport. Using a MacBook Pro 15" to view enwiki/Obama, the
"Languages" sidebar section in which a badge icon would be used,
is outside the viewport (shows until the first of three items of
the "In other projects" sidebar section instead).

On larger devices where it does fit the vieweport, it would still an
area where the user's primary focus usually is not at during the page
rendering. Either way, on fast connections it will still load during
the first paint. And on slower connections, it allows the layout
and text to render sooner (not blocking on the icons), with the
icons still rendering before page completes loading.

Bug: T121730
Bug: T127328
Change-Id: Idf12a20271d45ec053f264c06dbd3038c0e1fb0f