Remove misleading version number 1.0 from extension.json
This version number does not have any meaning (never had). It's
misleading, if not wrong. All code WMDE maintains is deployed weekly.
Branches are cut at a fixed time. Nobody ever assigned version
numbers to these releases.

The number is not used in communication. The only place where it
appears is Special:Version, but is useless there. Users can't use
it for anything. I talked to all relevant parties (product manager,
communication) and nobody ever used this version number. Because of
this, this is a purely technical task to get rid of unnecessary
technical dept.

Bug: T213066
Change-Id: I0bd452a0dd11f355950e6714c8ce4ac41b289546