Handle missing items in WikibaseClientSiteLinksForItemHandler
If the target value of a “topic’s main category” or “category related to
list” statement does not exist, that’s highly unusual, but we shouldn’t
crash in that case.

Also, fix a typo in another test case description that I just noticed.

The PhanTypeMismatchReturnSuperType suppression is also removed, because
apparently Phan no longer complains on this line, even though we still
return a narrower type (Item instead of EntityDocument).

Bug: T291953
Change-Id: Ia3c2fb5d0d6c5c382f3c67250541fd404a9950e7